Adds getOriginOfCurrentTab selector
